// src/types/news.ts
export interface ArticleSource {
  id: string | null;
  name: string;
}

export interface Article {
  source: ArticleSource;
  author: string | null;
  title: string;
  description: string | null;
  url: string;
  urlToImage: string | null;
  publishedAt: string;
  content: string | null;
}

export interface TopHeadlinesResponse {
  status: 'ok' | 'error';
  totalResults: number;
  articles: Article[];
}